fix: parse float instead of ints when getting quantity from req

+++ src/backend/bitcoinService/server.js
@@ -34,7 +34,7 @@ app.use(function (req, res, next) {
 });
 
 app.get("/", (req, res) => {
-  const quantity = parseInt(req.query["quantity"]);
+  const quantity = parseFloat(req.query["quantity"]);
 
   const currentBitcoinPrice =
     Math.round(Math.random() * 10000.75 + 50000 * 100) / 100;
